requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.1.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pywin32==307
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
